tests/frank-marco/inprogress/contact_save.html

<!doctype html>
<title>PhoneGap Capture test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

test(function() {
  assert_true(typeof navigator.contacts !== "undefined");
}, "Contacts (navigator.contacts) should exist");

var contact;
test(function() {
  contact = new Contact("a", "b", new ContactName("a", "b", "c", "d", "e", "f"), "c", [], [], [], [], [], "f", "i", [], [], []);  
  assert_true(typeof contact !== "undefined");  

  test_Contact(contact);
    
}, "should be able to create a Contact instance");

function test_Contact(contact) {
  test(function() {
  	assert_true(typeof contact.save !== "undefined");
  	assert_true(typeof contact.save === "function");
  }, "Contact.save should be a function");

  test(function() {
  	assert_true(typeof contact.remove !== "undefined");
  	assert_true(typeof contact.remove === "function");
  }, "Contact.remove should be a function");
  
  test(function() {
  	assert_true(contact.id === "a");
  }, "Contact.id should be 'a'");
  
  test(function() {
  	assert_true(contact.displayName === "b");
  }, "Contact.displayName should be 'b'");

  test(function() {
  	assert_true(contact.name.formatted === "a");
  }, "Contact.name.formatted should be 'a'");
  
  test(function() {
  	assert_true(contact.nickname === "c");
  }, "Contact.nickname should be 'c'");

  test(function() {
  	assert_true(typeof contact.phoneNumbers !== "undefined");
  }, "Contact.phoneNumbers should be exist");
  
  test(function() {
  	assert_true(typeof contact.emails !== "undefined");
  }, "Contact.emails should be exist");
  
  test(function() {
  	assert_true(typeof contact.addresses !== "undefined");
  }, "Contact.addresses should be exist");
  
  test(function() {
  	assert_true(typeof contact.ims !== "undefined");
  }, "Contact.ims should be exist");

  test(function() {
  	assert_true(typeof contact.organizations !== "undefined");
  }, "Contact.organizations should be exist");

  test(function() {
  	assert_true(contact.birthday === "f");
  }, "Contact.birthday should be 'f'");

  test(function() {
  	assert_true(contact.note === "i");
  }, "Contact.birthday should be 'i'");
  
  test(function() {
  	assert_true(typeof contact.photos !== "undefined");
  }, "Contact.photos should be exist");

  test(function() {
  	assert_true(typeof contact.categories !== "undefined");
  }, "Contact.categories should be exist");

  test(function() {
  	assert_true(typeof contact.urls !== "undefined");
  }, "Contact.urls should be exist");
}

var contactName;
test(function() {
  contactName = new ContactName("Dr. First Last Jr.", "Last", "First", "Middle", "Dr.", "Jr.");
  assert_true(typeof contactName !== "undefined");  

  test_ContactName(contactName);
  
}, "should be able to create a ContactName instance");

function test_ContactName(contactName) {
  test(function() {
  	assert_true(contactName.formatted === "Dr. First Last Jr.");
  }, "ContactName.formatted should be 'Dr. First Last Jr.'");

  test(function() {
  	assert_true(contactName.familyName === "Last");
  }, "ContactName.familyName should be 'Last'");

  test(function() {
  	assert_true(contactName.givenName === "First");
  }, "ContactName.givenName should be 'First'");

  test(function() {
  	assert_true(contactName.middleName === "Middle");
  }, "ContactName.middleName should be 'Middle'");
  
  test(function() {
  	assert_true(contactName.honorificPrefix === "Dr.");
  }, "ContactName.honorificPrefix should be 'Dr.'");

  test(function() {
  	assert_true(contactName.honorificSuffix === "Jr.");
  }, "ContactName.honorificSuffix should be 'Jr.'");
}

var contactField;
test(function() {
  contactField = new ContactField("home", "8005551212", true);  
  assert_true(typeof contactField !== "undefined");

  test_ContactField(contactField);
  
  var field = new ContactField(12345678, 12345678, true);

  test(function() {
  	assert_true(field.type === "12345678");
  }, "ContactField.type should be converted to the string '12345678'");

  test(function() {
  	assert_true(field.value === "12345678");
  }, "ContactField.value should be converted to the string '12345678'");

}, "should be able to create a ContactField instance");

function test_ContactField(contactField) {
  test(function() {
  	assert_true(contactField.type === "home");
  }, "ContactField.type should be 'home'");

  test(function() {
  	assert_true(contactField.value === "8005551212");
  }, "ContactField.value should be '8005551212'");

  test(function() {
  	assert_true(contactField.pref === true);
  }, "ContactField.pref should be true");
}

var contactAddress;
test(function() {
  contactAddress = new ContactAddress(true, "home", "a","b","c","d","e","f");
  assert_true(typeof contactAddress !== "undefined");

  test_ContactAddress(contactAddress);

}, "should be able to create a ContactAddress instance");

function test_ContactAddress(contactAddress) {
  test(function() {
  	assert_true(contactAddress.pref === true);
  }, "ContactAddress.pref should be true");
      
  test(function() {
  	assert_true(contactAddress.type === "home");
  }, "ContactAddress.type should be 'home'");

  test(function() {
  	assert_true(contactAddress.formatted === "a");
  }, "ContactAddress.formatted should be 'a'");

  test(function() {
  	assert_true(contactAddress.streetAddress === "b");
  }, "ContactAddress.streetAddress should be 'b'");

  test(function() {
  	assert_true(contactAddress.locality === "c");
  }, "ContactAddress.locality should be 'c'");

  test(function() {
  	assert_true(contactAddress.region === "d");
  }, "ContactAddress.region should be 'd'");

  test(function() {
  	assert_true(contactAddress.postalCode === "e");
  }, "ContactAddress.postalCode should be 'e'");

  test(function() {
  	assert_true(contactAddress.country === "f");
  }, "ContactAddress.country should be 'f'");  
}

var contactOrg;
test(function() {
  contactOrg = new ContactOrganization(true, "home", "a","b","c","d","e","f","g");
  assert_true(typeof contactOrg !== "undefined");
  
  test_ContactOrganization(contactOrg);
  
}, "should be able to create a ContactOrganization instance");

function test_ContactOrganization(contactOrg) {

  test(function() {
  	assert_true(contactOrg.pref === true);
  }, "ContactOrg.pref should be true");
      
  test(function() {
  	assert_true(contactOrg.type === "home");
  }, "ContactOrg.type should be 'home'");

  test(function() {
  	assert_true(contactOrg.name === "a");
  }, "ContactOrg.name should be 'a'");

  test(function() {
  	assert_true(contactOrg.department === "b");
  }, "ContactOrg.department should be 'b'");

  test(function() {
  	assert_true(contactOrg.title === "c");
  }, "ContactOrg.title should be 'c'");
}

test(function() {
  var contactFindOptions = new ContactFindOptions("a", true, "b");
  assert_true(typeof contactFindOptions !== "undefined");

  test(function() {
  	assert_true(contactFindOptions.filter === 'a');
  }, "ContactFindOptions.filter should be 'a'");
      
  test(function() {
  	assert_true(contactFindOptions.multiple === true);
  }, "ContactFindOptions.multiple should be true");

}, "should be able to create a ContactFindOptions instance");

test(function() {
  var contact = new Contact();
  assert_true(typeof contact.clone === "function");
  
}, "Contact.clone should be function");

test(function() {
  var contact = new Contact();
  contact.id=1;
  contact.displayName="Test Name";
  contact.nickname="Testy";
  contact.gender="male";
  contact.note="note to be cloned";
  contact.name = new ContactName("Mr. Test Name");

  var clonedContact = contact.clone();
  assert_true(typeof clonedContact !== "undefined");

  test(function() {
  	assert_true(contact.id === 1);
  }, "Contact.id should be 1");
  
  test(function() {
  	assert_true(clonedContact.id === null);
  }, "ClonedContact.id should be null");

  test(function() {
  	assert_true(clonedContact.displayName === contact.displayName);
  }, "ClonedContact.displayName should be contact.displayName");

  test(function() {
  	assert_true(clonedContact.nickname === contact.nickname);
  }, "ClonedContact.nickname should be contact.nickname");

  test(function() {
  	assert_true(clonedContact.gender === contact.gender);
  }, "ClonedContact.gender should be contact.gender");

  test(function() {
  	assert_true(clonedContact.note === contact.note);
  }, "ClonedContact.note should be contact.note");

  test(function() {
  	assert_true(clonedContact.name.formatted === contact.name.formatted);
  }, "ClonedContact.name.formatted should be contact.name.formatted");

  test(function() {
  	assert_true(clonedContact.connected === contact.connected);
  }, "ClonedContact.connected should be contact.connected");
  
}, "Contact.clone function should make deep copy of Contact Object");

</script>
